# ruoyi-agent/docs/agent-v2-tools.openapi.yaml
openapi: 3.0.3
info:
  title: RuoYi Agent V2 Tool Gateway
  version: 2.0.0
  description: Dify Supervisor 调用的企业业务工具。所有请求必须同时携带服务密钥和本次运行令牌。
servers:
  - url: https://YOUR_PUBLIC_RUOYI_HOST
security:
  - ToolGatewayKey: []
paths:
  /internal/agent-tools/query-business-data:
    post:
      operationId: query_business_data
      summary: 查询经过权限校验的企业业务数据
      description: 第一版支持物料按分类、物料类型或状态聚合计数。禁止传入 SQL。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
            examples:
              materialCount:
                value:
                  runId: 1
                  runToken: '{{run_token}}'
                  idempotencyKey: '{{run_id}}-material-category-count'
                  input:
                    domain: MATERIAL
                    operation: COUNT
                    groupBy: CATEGORY
                    goal: 分析物料分类数量
                    filters: {}
      responses:
        '200':
          description: 标准工具结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResult'
  /internal/agent-tools/analyze-dataset:
    post:
      operationId: analyze_dataset
      summary: 对已查询的数据集执行确定性统计
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeRequest'
            examples:
              statistics:
                value:
                  runId: 1
                  runToken: '{{run_token}}'
                  idempotencyKey: '{{run_id}}-analyze-latest'
                  input:
                    datasetId: 12
                    valueField: materialCount
      responses:
        '200':
          description: 标准工具结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResult'
  /internal/agent-tools/render-chart:
    post:
      operationId: render_chart
      summary: 将已有数据集渲染为安全语义图表
      description: datasetId 为空时，Spring 自动使用当前会话最近的数据集。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenderChartRequest'
            examples:
              chart:
                value:
                  runId: 1
                  runToken: '{{run_token}}'
                  idempotencyKey: '{{run_id}}-chart-bar'
                  input:
                    datasetId: 12
                    chartType: bar
                    categoryField: dimensionName
                    valueField: materialCount
                    title: 物料分类数量
      responses:
        '200':
          description: 标准工具结果，artifacts 中包含图表协议
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResult'
  /internal/agent-tools/request-clarification:
    post:
      operationId: request_clarification
      summary: 在关键参数缺失时生成一个明确的追问
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClarificationRequest'
            examples:
              clarification:
                value:
                  runId: 1
                  runToken: '{{run_token}}'
                  idempotencyKey: '{{run_id}}-clarify-1'
                  input:
                    question: 你希望按物料分类、物料类型还是状态统计？
      responses:
        '200':
          description: 标准工具结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResult'
components:
  securitySchemes:
    ToolGatewayKey:
      type: apiKey
      in: header
      name: X-Agent-Tool-Key
  schemas:
    ToolIdentity:
      type: object
      required: [runId, runToken, idempotencyKey]
      properties:
        runId:
          type: integer
          format: int64
          description: 使用 Dify 输入变量 run_id
        runToken:
          type: string
          description: 使用 Dify 输入变量 run_token，不得自行生成
        idempotencyKey:
          type: string
          description: 同一次逻辑工具调用必须稳定且唯一
    QueryRequest:
      allOf:
        - $ref: '#/components/schemas/ToolIdentity'
        - type: object
          required: [input]
          properties:
            input:
              type: object
              required: [domain, operation, groupBy]
              properties:
                domain:
                  type: string
                  enum: [MATERIAL]
                operation:
                  type: string
                  enum: [COUNT]
                groupBy:
                  type: string
                  enum: [CATEGORY, MATERIAL_TYPE, STATUS]
                goal:
                  type: string
                filters:
                  type: object
                  properties:
                    materialType:
                      type: string
                      enum: [RAW, SEMI_FINISHED, FINISHED, AUXILIARY, PACKAGE]
                    categoryId:
                      type: integer
                      format: int64
                    status:
                      type: string
    AnalyzeRequest:
      allOf:
        - $ref: '#/components/schemas/ToolIdentity'
        - type: object
          required: [input]
          properties:
            input:
              type: object
              properties:
                datasetId:
                  type: integer
                  format: int64
                  description: 为空时自动使用 last_dataset_id
                valueField:
                  type: string
    RenderChartRequest:
      allOf:
        - $ref: '#/components/schemas/ToolIdentity'
        - type: object
          required: [input]
          properties:
            input:
              type: object
              required: [chartType]
              properties:
                datasetId:
                  type: integer
                  format: int64
                  description: 为空时自动使用 last_dataset_id
                chartType:
                  type: string
                  enum: [bar, line, pie]
                categoryField:
                  type: string
                valueField:
                  type: string
                title:
                  type: string
    ClarificationRequest:
      allOf:
        - $ref: '#/components/schemas/ToolIdentity'
        - type: object
          required: [input]
          properties:
            input:
              type: object
              required: [question]
              properties:
                question:
                  type: string
    ToolResult:
      type: object
      required: [status, data, artifacts, citations, requiresConfirmation, nextActions]
      properties:
        status:
          type: string
          enum: [SUCCEEDED, FAILED, NEEDS_CLARIFICATION, WAITING_CONFIRMATION]
        summary:
          type: string
        data:
          type: object
          additionalProperties: true
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/Artifact'
        citations:
          type: array
          items:
            type: object
            additionalProperties: true
        requiresConfirmation:
          type: boolean
        nextActions:
          type: array
          items:
            type: string
        error:
          type: object
          nullable: true
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
    Artifact:
      type: object
      required: [type, version, title, payload]
      properties:
        artifactId:
          type: integer
          format: int64
        type:
          type: string
          enum: [CHART, TABLE, CODE, FILE, DOCUMENT]
        version:
          type: string
          example: '2.0'
        title:
          type: string
        datasetId:
          type: integer
          format: int64
        payload:
          type: object
          additionalProperties: true
        fileId:
          type: integer
          format: int64
        mimeType:
          type: string
        previewUrl:
          type: string
        downloadUrl:
          type: string